+++ src/pages/references.js
@@ -1,14 +1,17 @@
-import React, { useEffect, useState } from 'react'
+import React from 'react'
+import { Router } from '@reach/router'
+import Reference from './references/reference'
+import ReferencesPage from './references/references'
 
-import { getAllReferences, Reference } from '../models/Reference'
-
-const ReferencesPage = () => {
-  useEffect(() => {
-    const refs = getAllReferences().then(res =>
-      res.forEach(doc => console.log(doc.data()))
-    )
-  }, [])
-  return <div />
+const References = () => {
+  return (
+    <div>
+      <Router>
+        <ReferencesPage path='/references' />
+        <Reference path='/references/:reference' />
+      </Router>
+    </div>
+  )
 }
 
-export default ReferencesPage
+export default References
