+++ src/pages/whitepaper.js
@@ -2,14 +2,14 @@ import React from 'react'
 
 import { database } from '../../firebase'
 
-const documents = database
-  .collection('references')
-  .get()
-  .then(querySnapshot => {
-    querySnapshot.forEach(doc => {
-      console.log(`${doc.id} => ${doc.data()}`)
-    })
-  })
+// const documents = database
+//   .collection('references')
+//   .get()
+//   .then(querySnapshot => {
+//     querySnapshot.forEach(doc => {
+//       console.log(`${doc.id} => ${doc.data()}`)
+//     })
+//   })
 
 const WhitepaperPage = () => {
   return <div />
